test(dedup): basis to figure out types and API
We now provide test-fixtures based on actual package.json's, both
in the new (flat) as well as the (previous) nested format.

That way we should be able to test all the nitty-gritty details
and get it to work even before it hits the command-line. The latter
should just be a formality after all.