imp(dedup): ignore existing symlinks
However, next step is to assure the symlink is already pointing to the
correct destination. If that is not the case, we would emit an instruction.

It doesn't seem to be too comfortable to read a symbolic link though, but it's
definitely possible. At least it's clear that doing this is highly platform
dependent, and letting people work for it makes this obvious.

Thinking about it: the only reason the symlink would not be valid is if
we wanted to change it from absolute to relative, if it was broken, or
if for some reason the package.json contained inside does not match
the repositories directory.